Allow multi-file analysis of C programs (#26)
* Add import mechanism to Gillian-C (#24)

* Change CLI to take multiple input files

* Change GIL imports to be enclosed in quotes (#24)

* Add linking and merging step to Gillian-C (#24)

* Ignore GIL-defined functions in linking step (#24)

* Correctly initialise i__global_env (#24)

* Ignore generated GIL and JSIL files (#24)

* Add option to specify .h search dirs (#24)

* Add option to specify source dir (#24)

This would be a directory that, for example, contains the .c files of
the library for which symbolic tests are written and therefore need to
be linked against. It is searched recursively.

* Correctly handle static functions in linker (#24)

* Add options to suppress linker errors (#24)

* Mangle variables with reserved names (#24)

* Correctly print float values when writing GIL files (#24)

'Fmt.float' uses the '%g' format type, which attempts to convert the
number to the shortest possible representation. This led to instances
of numbers being written in scientific notation, e.g. 2.24682e+09,
which cannot be parsed by the GIL parser.

* Suppress linker errors in CI tests (#24)

* Don't hardcode ASSUME and ASSERT (#24)

* Add C examples using externally-defined funcs and vars (#24)

* Display trailing dot for float values (#24)

* fix types

* Fix Gillian-C imports for verification and bi-abduction (#25)

* Don't generate duplicate predicates for the same struct

* Add bi-abduction example

* Explicitly declare which GIL imports need to be verified

'eprog_to_prog' by default marks the specification of every imported
procedure as not to be verified. Therefore, add a way to distinguish
between the different kinds of imports, i.e. runtime files which
should not be verified, and GIL files corresponding to user files,
which should.

* Include bi-abduction specs from imported programs

* moving stateful components out

* Refactor eprog_to_prog (#27)

Refactor duplicate code into separate functions, reduce the size of
overlong comments and simplify some of the import-resolving logic. Add
the missing 'else' expression when the current import to be resolved
has been seen already.

Co-authored-by: Alexis Marinoiu <34110130+alexismarinoiu@users.noreply.github.com>