[WIP] Better reporting system (#22)
* Logging: nit, use `format` instead of `format4`

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: add @opam/uuidm

Next commit requires @opam/uuidm, this includes all the boilerplate
changes as a result of adding it.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: implement reports

Reports are custom logging units that contain extra information
alongside a logging message. They help preserve the hierarchy of logs
in as a tree structure.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: implement file reporter and interface

Reporters implement the logic for producing the logs in different
formats. Currently this implements a file reporter which reports logs
as before. It also provides support for other more complex reporters
to be implemented in the future such as a database reporter.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: log using reporters

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Fixing some documentation

Co-authored-by: Sacha Ayoun <sachaayoun@gmail.com>