Fix slowness when running bulk tests in Gillian-C (#34)
* Initialise CompCert only once per bulk run

Previously, the include headers were added on every call to
parse_and_compile_file, which meant the list of arguments passed to
the C preprocessor kept on growing. This sometimes resulted
in errors during bulk testing of Collections-C.

* Cache compiled progs in both Gillian and Gillian-C

* Don't write imported GIL programs to file in Gillian-C

This is until we figure out a more systematic way to handle
intermediate GIL files and their storage. It makes the bulk tests run
a lot faster.