Track changes in lemmas
Like with predicates, if a lemmas changes, then all the lemmas that
are transitively using the lemma as well as all procedures that are
using these lemmas within their proof tactics must be marked as
changed. In addition, when a predicate changes, keep track of the
lemmas transitively using this predicate.

Needs to be tested more thoroughly. Maybe once we have a working WISL
example?