Add support for continuous verification (#39)
* Add writing and reading of file contents hashes

* Add functions to detect changed files

* Keep track of each procedure's source file path

* Add call graph definition and related functions

These include functions to serialize and deserialize a call graph as
JSON.

* Add basic rules to determine which procs to re-verify

These make use of the stored call graph from the previous analysis run
and the information about which source files have been changed, added,
or deleted.

* Define verification results type and its related JSON functions

* Display both current and previous analysis results

* Add CLI flag to enable incremental analysis

* Keep track of .c file contents as opposed to preprocessed ones

The latter exclude comments, making it impossible to determine if any
of the specs have changed. Need to figure out how to handle changes in
.h files.

* Add directive in GIL to specify procs with no source paths

This is intended to mark functions that are added during compilation
to GIL but which do not correspond to real functions from the source
file (e.g. 'i__initialize_genv' in Gillian-C).

* Add ability to mark files and procs in GIL as internal

This can be done by either adding "internal;" at the beginning of the
GIL file or by adding "@internal" before a specific GIL proc. Use this
instead of source path to decide which procs to add to the call graph.

* Make JSON functions be like the ones from 'ppx_deriving_yojson'

Use the plugin to derive serialization and deserialization functions
wherever possible. May switch to 'ppx_yojson_conv' later as this
supports hash tables.

* Remove empty files

* Move verification results logic to Verifier.ml

In addition, move CallGraph and VerificationResults to the results
subdir in order to avoid creating a dependency cycle between engine
and results.

* Add ability to reset state of the results globals

* Handle deleted procedures properly

* Extract the included headers of each source file in Gillian-C

Use the various C preprocessor '-M' options in order to get all the
files that were included following the preprocessing step, dump these
to a temporary .deps file, and parse and extract the header paths from
this.

* Handle changes in header files

For now, if a header file changes betwen analyses, mark every source
file that includes it as changed.

* Add several tests for file-diff-based incremental verification

More will follow.

* Determine which predicates each procedure depends on

These can be located within the spec of the proc or within proof
tactics inside the body.

* Keep track of dependencies between preds and procs in call graph

* Track changes in predicates

Extend the Pred.t record to include a source path and a flag
indicating whether it is internally-defined, like in Proc.t.

Extend the heuristics in changeTracker.ml to also take into account
changed predicates. If a predicate changes, then all the predicates
that are transitively using it as well as all procedures that use
these predicates (either in their specs or within proof tactics) must
be marked as changed.

* Add test for changing a predicate definition

* Rename module

* Rename the internal file directive

* Switch to using a reduce visitor to extract predicates used

* root dirs for dune

* factorize common cli args and add result-dir arg to cli

* replace sh with dune script

* simpler dune

* a tiny bit more abstraction

* Remove remaining test bash scripts

* Extend lemma def to include source path and internal marker

* Keep track of lemma dependencies in call graph

* Track changes in lemmas

Like with predicates, if a lemmas changes, then all the lemmas that
are transitively using the lemma as well as all procedures that are
using these lemmas within their proof tactics must be marked as
changed. In addition, when a predicate changes, keep track of the
lemmas transitively using this predicate.

Needs to be tested more thoroughly. Maybe once we have a working WISL
example?

* Change Verifier.ml to only re-analyse changed lemmas

* Add fixme note for later

Co-authored-by: Sacha Ayoun <sachaayoun@gmail.com>