Fix call graph merging
Change the `CallGraph.merge` function to only add entries that were
not already included in the first graph, in effect passing the
responsibility to the caller to ensure the first graph was pruned
correctly. This is to fix the bug where a changed procedure calls a
non-changed procedure and therefore creates an empty node entry for
the callee that would then overwrite the original one.