Fix heuristic for getting transitive callers in verification
When a procedure changes in verification mode, its closest caller
(i.e. ancestor of the call graph) with a spec should be marked as a
dependent and therefore be re-verified. Use an adapted version of
`get_proc_callers` in both `get_changes` and `get_verif_changes` which
addresses this. In particular, we should now properly deal with the
following case (-> indicates a call):

A -> B -> B' -> C

where A and C are functions with specs and B and B' are non-internal
functions without specs. If C changes, then A should be re-verified.
The test `test_transitive_call` checks this.