De-couple compilation from execution during bulk testing
This is to be able to use the ChangeTracker in order to filter out the
tests that do not need to be re-run.

In addition, keeping everything within the same call graph will not
likely be useful, since it will be hard to ensure that procedures
defined in different test files with the same name (e.g.
`teardown_test()` in Collections-C) will not clash.