Re-configure runner to perform incremental decision during test
De-coupling compilation from execution meant that all the compiled
programs had to be kept in memory in order for them to be later
fetched during execution. The memory usage and subsequent garbage
collection becomes an issue with very large test suites such as
Test262.

Move towards an implementation where the tests are not simply
skipped, but have their summaries fetched and displayed to the user,
therefore making it seem like they were ran again.