Improve incremental ACT algorithm
In incremental mode, instead of re-checking every transitive caller of
a changed procedure, execute the procedure first and only check its
closest non-internal callers if its new derived spec has changed from
the previously stored one.

Compute hashes of specs in order to compare them. In practice, this is
not robust since the order in which the procedures are executed
affects the allocated locations. A better comparison would be to check
whether there is a substitution between variables in the two specs.