Add support for continuous symbolic testing (#55)
* Add support for incremental symbolic testing

If the `main` procedure is in the set of transitive callers of
changed procedures, then re-run the test, otherwise don't.

* Add incremental test for case when everything is unchanged

* Cache header dependencies files during compilation

* Keep track of combined call graph during bulk symbolic tests

Add a prefix consisting of the test file name to each `main` function
in order to keep them unique within the graph.

* Create interfaces for SourceFiles and ChangeTracker

* Tidy up Gilgen interface

* Remove use of allocators during compilation

* De-couple compilation from execution during bulk testing

This is to be able to use the ChangeTracker in order to filter out the
tests that do not need to be re-run.

In addition, keeping everything within the same call graph will not
likely be useful, since it will be hard to ensure that procedures
defined in different test files with the same name (e.g.
`teardown_test()` in Collections-C) will not clash.

* Add reading and writing of bulk call graphs and source files

* Determine the minimum number of tests to re-run

* Rename interpreter in `Runner.ml`

It is not always a concrete interpreter.

* Cache computed hashes of source files

This speeds up compilation time considerably.

* Add missing lemma processing step in `gil_parsing`

* Add interfaces for modules inside GIL_Parser

This makes it clearer to see what is called externally. Remove the
`parse_line_numbers` function as this does not seem to be used.

* Update README

Badges are cool.

* Fix GIL file extension check

* Fix formatting

* Fix bug where all previously compiled GIL progs are included

* Re-configure runner to perform incremental decision during test

De-coupling compilation from execution meant that all the compiled
programs had to be kept in memory in order for them to be later
fetched during execution. The memory usage and subsequent garbage
collection becomes an issue with very large test suites such as
Test262.

Move towards an implementation where the tests are not simply
skipped, but have their summaries fetched and displayed to the user,
therefore making it seem like they were ran again.

Co-authored-by: Sacha Ayoun <sachaayoun@gmail.com>