Logging: fix symb exec report branching (#67)
* Logging: use unique ids for phases

Ensure correctness of closing the phases in a LIFO order by using
unique identifiers instead of checking on the phase type. Opening a
phase produces a token that must be used in closing it.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: Report.id = (pid * uuid)

Pair the unique id of reports with the current pid. Fixes issue when
forking would produce reports with duplicate uuids because `Uuidm`
generates a deterministic sequence from the same random seed.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>

* Logging: Database Reporter parallel support

The database file must be locked since multiple processes could write
to it in parallel mode.

Signed-off-by: Radu Lacraru <radu.lacraru97@gmail.com>