Fixes for WISL, debugging, loop invariants (#341)
* Add fake roots for proc-submaps

This is a bit of a hack, it allows us to "bootstrap" proc submaps since you need a previous in order to step in SEDAP

* Squash Node_updated events from GIL lifter when TL lifter is present

* Add `hidden` and `display_name` fields to Procs for the debugger

* Add type assertion to WISL

* Fail instead of warn when failing to make MP for post

* Bump linol version

* Tweak `Logging.with_id` to not leak backtraces to stdout

* Perform "heartbeat" after each SMT command

If the solver dies, it won't be detected until the next time you use the relevant channels. This way, we'll know that SMT died right away (plus it's a nice sanity check anyways).

* Replace SLL_iterative with iterative lab functions, add to quicktests

* Consider errors during `testify` as analysis failures, not compilation errors

* Handle "lst-cat of lst-rev" preprocessing edge case

* CI: don't restore stale caches