core/smt: record BV variant for each BV expr arg
Consider the following GIL pseudocode:

```
assert(typeOf(#lvar_1) = List)
bvnot(Bitvector(l-nth(#lvar_1, 1i), 64i) :)
```

No bitvector types will be defined in the SMT context for this snippet,
because bvnot has a width of None and the LVar has a type of List. There
is information about the first element of the list, but it's encoded in
the op itself as a bv_arg, and not in gamma.

To resolve this, we now add bv_arg widths to the defined_bv_variants
when encoding bitvector expressions.